requirements/base.txt
certifi==2020.4.5.2
chardet==3.0.4
Django==2.2.13
django-markdown-app==0.9.6
django-nopassword==4.0.0
django-paypal==1.0.0
docopt==0.6.2
docutils==0.16
idna==2.9
Jinja2==2.11.2
mailchimp3==3.0.14
Markdown==3.2.2
psycopg2-binary==2.8.5
Pillow==7.1.2
pytz==2020.1
requests==2.23.0
six==1.15.0
Unidecode==1.1.1
urllib3==1.25.9
